# services/horizon/internal/db2/history: Improve performance of effects query