--- a/services/horizon/internal/actions/effects.go
+++ b/services/horizon/internal/actions/effects.go
@@ -95,25 +95,20 @@ func (handler GetEffectsHandler) GetResourcePage(w HeaderWriter, r *http.Request
 }
 
 func loadEffectRecords(ctx context.Context, hq *history.Q, qp EffectsQuery, pq db2.PageQuery) ([]history.Effect, error) {
-	effects := hq.Effects()
-
 	switch {
 	case qp.AccountID != "":
-		effects.ForAccount(ctx, qp.AccountID)
+		return hq.EffectsForAccount(ctx, qp.AccountID, pq)
 	case qp.LiquidityPoolID != "":
-		effects.ForLiquidityPool(ctx, pq, qp.LiquidityPoolID)
+		return hq.EffectsForLiquidityPool(ctx, qp.LiquidityPoolID, pq)
 	case qp.OperationID > 0:
-		effects.ForOperation(int64(qp.OperationID))
+		return hq.EffectsForOperation(ctx, int64(qp.OperationID), pq)
 	case qp.LedgerID > 0:
-		effects.ForLedger(ctx, int32(qp.LedgerID))
+		return hq.EffectsForLedger(ctx, int32(qp.LedgerID), pq)
 	case qp.TxHash != "":
-		effects.ForTransaction(ctx, qp.TxHash)
+		return hq.EffectsForTransaction(ctx, qp.TxHash, pq)
+	default:
+		return hq.Effects(ctx, pq)
 	}
-
-	var result []history.Effect
-	err := effects.Page(pq).Select(ctx, &result)
-
-	return result, err
 }
 
 func loadEffectLedgers(ctx context.Context, hq *history.Q, effects []history.Effect) (map[int32]history.Ledger, error) {