--- a/services/horizon/internal/db2/history/effect_batch_insert_builder_test.go
+++ b/services/horizon/internal/db2/history/effect_batch_insert_builder_test.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/guregu/null"
 
+	"github.com/stellar/go/services/horizon/internal/db2"
 	"github.com/stellar/go/services/horizon/internal/test"
 	"github.com/stellar/go/toid"
 )
@@ -42,8 +43,12 @@ func TestAddEffect(t *testing.T) {
 	tt.Assert.NoError(builder.Exec(tt.Ctx, q))
 	tt.Assert.NoError(q.Commit())
 
-	effects := []Effect{}
-	tt.Assert.NoError(q.Effects().Select(tt.Ctx, &effects))
+	effects, err := q.Effects(tt.Ctx, db2.PageQuery{
+		Cursor: "0-0",
+		Order:  "asc",
+		Limit:  200,
+	})
+	tt.Require.NoError(err)
 	tt.Assert.Len(effects, 1)
 
 	effect := effects[0]